[feat] 금융 상품 요약 조회 및 계산 기능 구현 (#81)
* feat - Implement summary retrieval for financial products and refactor response DTOs

* change assert code - jsonPath

* feat - Enhance financial product summary retrieval with detailed sorting and additional fields

* feat - Add methods to calculate expiration date, progress, and remaining days for financial products

* feat - Introduce LocalDateProvider interface and SystemDateTimeProvider implementation for date management

* feat - Add balance calculation and summary retrieval for savings financial products

* add method docs

* feat - Refactor financial product summary retrieval to use a builder method for response creation and add CashInvestment implementation

* feat - Extract comparator logic for financial product summary sorting into a separate method

* delete comment

* feat - Update financial product handling to set maximum expiration date for cash investments and adjust method parameters for summary response

* feat - Refactor expiration date calculation for investment types to use a strategy pattern

* feat - Add FinancialProductDataProvider for creating test instances and update tests to use it

* add test - deposit getProgressByLocalDate

* feat - Refactor progress calculation for investment types to use a strategy pattern

* feat - Rename calculateExpirationDate method to calculate in ExpirationCalculator interface and update InvestmentType to reflect the change

* feat - Introduce PeriodStrategy and refactor expiration date and progress calculations for investment types

* feat - Update progress and remaining days calculations for cash and savings products, and add corresponding tests

* feat - Add tests for progress calculation of savings products based on various date scenarios

* add test - getBalanaceByLocalDate

* feat - Implement balance calculation strategies for financial products and refactor related methods

* refactor - Simplify balance calculation by removing redundant method and updating balance retrieval in summary response

* refactor - Use ChronoUnit directly for calculating elapsed months in balance policy

* inline variable - FinancialProductSummaryResponse list

* feat - Add DetailedFinancialProductResponse and update product detail retrieval in service and controller

* introduce varaible - today

* refactor - Update tests to use InvestmentType instead of ProductType

* refactor - Change total investment type from int to BigDecimal across investment classes

* refactor - Change return type of interest calculation methods from int to BigDecimal across investment classes

* refactor - Replace ToIntFunction with UnaryOperator for rounding in Investment interface

* fix typo field - createdAt

* add comparator - id

* fix test name

* add test

* add test - BalancePolicy

* add PaymentDay field to FinancialProduct and validation logic

* add investmentType validate logic

* add PaymentDay validation annotation and integrate into FinancialProduct

* add PaymentDay logic to BalancePolicy and corresponding tests

* fix test

+++ src/main/java/co/invest72/investment/application/CalculateInvestment.java
@@ -1,5 +1,6 @@
 package co.invest72.investment.application;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,10 +19,10 @@ public class CalculateInvestment {
 	public CalculateInvestmentResponse calculate(Investment investment) {
 		List<MonthlyInvestmentResult> monthlyDetails = getMonthlyInvestmentResults(investment);
 		List<YearlyInvestmentResult> yearlyDetails = getYearlyInvestmentResults(investment);
-		int totalInvestment = investment.getTotalInvestment();
-		int totalInterest = investment.getTotalInterest();
-		int totalTax = investment.getTotalTax();
-		int totalProfit = investment.getTotalProfit();
+		BigDecimal totalInvestment = investment.getTotalInvestment();
+		BigDecimal totalInterest = investment.getTotalInterest();
+		BigDecimal totalTax = investment.getTotalTax();
+		BigDecimal totalProfit = investment.getTotalProfit();
 		String taxType = investment.getTaxType();
 		String taxPercent = taxFormatter.format(investment.getTaxRate());
 
@@ -39,10 +40,10 @@ public class CalculateInvestment {
 
 	public CalculateMonthlyInvestmentResponse calMonthlyInvestment(Investment investment) {
 		List<MonthlyInvestmentResult> result = getMonthlyInvestmentResults(investment);
-		int totalInvestment = investment.getTotalInvestment();
-		int totalInterest = investment.getTotalInterest();
-		int totalTax = investment.getTotalTax();
-		int totalProfit = investment.getTotalProfit();
+		BigDecimal totalInvestment = investment.getTotalInvestment();
+		BigDecimal totalInterest = investment.getTotalInterest();
+		BigDecimal totalTax = investment.getTotalTax();
+		BigDecimal totalProfit = investment.getTotalProfit();
 		String taxType = investment.getTaxType();
 		String taxPercent = taxFormatter.format(investment.getTaxRate());
 		return CalculateMonthlyInvestmentResponse.builder()
@@ -72,10 +73,10 @@ public class CalculateInvestment {
 
 	public CalculateYearlyInvestmentResponse calYearlyInvestment(Investment investment) {
 		List<YearlyInvestmentResult> details = getYearlyInvestmentResults(investment);
-		int totalInvestment = investment.getTotalInvestment();
-		int totalInterest = investment.getTotalInterest();
-		int totalTax = investment.getTotalTax();
-		int totalProfit = investment.getTotalProfit();
+		BigDecimal totalInvestment = investment.getTotalInvestment();
+		BigDecimal totalInterest = investment.getTotalInterest();
+		BigDecimal totalTax = investment.getTotalTax();
+		BigDecimal totalProfit = investment.getTotalProfit();
 		String taxType = investment.getTaxType();
 		String taxPercent = taxFormatter.format(investment.getTaxRate());
 		return CalculateYearlyInvestmentResponse.builder()
@@ -94,9 +95,9 @@ public class CalculateInvestment {
 
 		int years = (investment.getFinalMonth() - 1) / 12 + 1;
 		for (int year = 1; year <= years; year++) {
-			int principal = investment.getPrincipalForYear(year);
-			int interest = investment.getInterestForYear(year);
-			int profit = investment.getProfitForYear(year);
+			BigDecimal principal = investment.getPrincipalForYear(year);
+			BigDecimal interest = investment.getInterestForYear(year);
+			BigDecimal profit = investment.getProfitForYear(year);
 			details.add(new YearlyInvestmentResult(year, principal, interest, profit));
 		}
 		return details;
